+++ clients/margin_trading/src/binance_sdk_margin_trading/rest_api/api/market_data_api.py
@@ -23,6 +23,7 @@ from ..models import GetAllMarginAssetsResponse
 from ..models import GetDelistScheduleResponse
 from ..models import GetLimitPricePairsResponse
 from ..models import GetListScheduleResponse
+from ..models import GetMarginAssetRiskBasedLiquidationRatioResponse
 from ..models import QueryIsolatedMarginTierDataResponse
 from ..models import QueryLiabilityCoinLeverageBracketInCrossMarginProModeResponse
 from ..models import QueryMarginAvailableInventoryResponse
@@ -309,6 +310,42 @@ class MarketDataApi:
             body=body,
             time_unit=self._configuration.time_unit,
             response_model=GetListScheduleResponse,
+        )
+
+    def get_margin_asset_risk_based_liquidation_ratio(
+        self,
+    ) -> ApiResponse[GetMarginAssetRiskBasedLiquidationRatioResponse]:
+        """
+                Get Margin Asset Risk-Based Liquidation Ratio (MARKET_DATA)
+                GET /sapi/v1/margin/risk-based-liquidation-ratio
+                https://developers.binance.com/docs/margin_trading/market-data/Get-Margin-Asset-Risk-Based-Liquidation-Ratio
+
+                Get Margin Asset Risk-Based Liquidation Ratio
+
+        Weight: 1
+
+                Args:
+
+                Returns:
+                    ApiResponse[GetMarginAssetRiskBasedLiquidationRatioResponse]
+
+                Raises:
+                    RequiredError: If a required parameter is missing.
+
+        """
+
+        body = None
+        payload = None
+
+        return send_request(
+            self._session,
+            self._configuration,
+            method="GET",
+            path="/sapi/v1/margin/risk-based-liquidation-ratio",
+            payload=payload,
+            body=body,
+            time_unit=self._configuration.time_unit,
+            response_model=GetMarginAssetRiskBasedLiquidationRatioResponse,
         )
 
     def query_isolated_margin_tier_data(