+++ clients/spot/src/binance_sdk_spot/websocket_api/api/account_api.py
@@ -200,7 +200,7 @@ class AccountApi:
 
             Args:
                     id (Optional[str] = None): Unique WebSocket request ID.
-                    from_id (Optional[int] = None): Aggregate trade ID to begin at
+                    from_id (Optional[int] = None): Trade ID to begin at
                     start_time (Optional[int] = None):
                     end_time (Optional[int] = None):
                     limit (Optional[int] = None): Default: 100; Maximum: 5000
@@ -512,7 +512,7 @@ class AccountApi:
                     order_id (Optional[int] = None): `orderId`or`origClientOrderId`mustbesent
                     start_time (Optional[int] = None):
                     end_time (Optional[int] = None):
-                    from_id (Optional[int] = None): Aggregate trade ID to begin at
+                    from_id (Optional[int] = None): Trade ID to begin at
                     limit (Optional[int] = None): Default: 100; Maximum: 5000
                     recv_window (Optional[float] = None): The value cannot be greater than `60000`. <br> Supports up to three decimal places of precision (e.g., 6000.346) so that microseconds may be specified.
 
@@ -669,7 +669,7 @@ class AccountApi:
                     order_id (Union[int, None]):
                     id (Optional[str] = None): Unique WebSocket request ID.
                     from_execution_id (Optional[int] = None):
-                    limit (Optional[int] = None): Default: 100; Maximum: 5000
+                    limit (Optional[int] = None): Default: 500; Maximum: 1000
                     recv_window (Optional[float] = None): The value cannot be greater than `60000`. <br> Supports up to three decimal places of precision (e.g., 6000.346) so that microseconds may be specified.
 
             Returns: