+++ clients/wallet/tests/unit/rest_api/test_asset_api.py
@@ -197,6 +197,7 @@ class TestAssetApi:
                     "divTime": 1563189166000,
                     "enInfo": "BHFT distribution",
                     "tranId": 2968885920,
+                    "direction": 1,
                 },
                 {
                     "id": 1631750237,
@@ -205,6 +206,7 @@ class TestAssetApi:
                     "divTime": 1563189165000,
                     "enInfo": "BHFT distribution",
                     "tranId": 2968885920,
+                    "direction": 1,
                 },
             ],
             "total": 2,
@@ -266,6 +268,7 @@ class TestAssetApi:
                     "divTime": 1563189166000,
                     "enInfo": "BHFT distribution",
                     "tranId": 2968885920,
+                    "direction": 1,
                 },
                 {
                     "id": 1631750237,
@@ -274,6 +277,7 @@ class TestAssetApi:
                     "divTime": 1563189165000,
                     "enInfo": "BHFT distribution",
                     "tranId": 2968885920,
+                    "direction": 1,
                 },
             ],
             "total": 2,